Windows: move BSD socket shims to netc
